rapiin controller

+++ application/models/Kamar_model.php
@@ -19,13 +19,23 @@ class Kamar_model extends CI_Model {
     }
 
     public function get_all_image() {
-        $query = $this->db->get('kamar_foto'); 
-        return $query->result_array();
+        $query = $this->db->get('kamar_foto');
+        $result = $query->result_array();
+        $foto = [];
+        foreach ($result as $row) {
+            $foto[$row['type_kamar']][] = $row;
+        }
+        return $foto;
     }
 
     public function get_all_spek() {
-        $query = $this->db->get('kamar_spek'); 
-        return $query->result_array();
+        $query = $this->db->get('kamar_spek');
+        $result = $query->result_array();
+        $spek = [];
+        foreach ($result as $row) {
+            $spek[$row['type_kamar']][] = $row;
+        }
+        return $spek;
     }
 
     public function get_image($id) {
@@ -43,13 +53,9 @@ class Kamar_model extends CI_Model {
     }
 
     public function get_foto_by_type_kamar($type_kamar) {
-        $this->db->select('foto');
         $this->db->where('type_kamar', $type_kamar);
         $query = $this->db->get('kamar_foto');
-        if ($query->num_rows() > 0) {
-            return $query->row_array();
-        }
-        return false;
+        return $query->result_array();
     }
 
     public function get_count_foto_by_type($type_kamar) {
@@ -62,21 +68,28 @@ class Kamar_model extends CI_Model {
         return $this->db->count_all_results('kamar_spek'); 
     }
 
-    public function delete_foto($id){
-        $this->db->where('id_foto', $id);
-        $this->db->delete('kamar_foto');
-
-        if ($this->db->affected_rows() > 0) {
-            return true; 
+    public function delete_foto($id) {
+        $foto = $this->db->get_where('kamar_foto', ['id_foto' => $id])->row_array();
+        if ($foto) {
+            $foto_path = './assets/img/upload/kamar/' . $foto['foto'];
+            if (file_exists($foto_path)) {
+                unlink($foto_path);
+            }
+            $this->db->where('id_foto', $id);
+            $this->db->delete('kamar_foto');
+            if ($this->db->affected_rows() > 0) {
+                return true; 
+            } else {
+                return false;
+            }
         } else {
             return false;
         }
     }
-
+    
     public function delete_spek($type_kamar){
         $this->db->where('type_kamar', $type_kamar);
         $this->db->delete('kamar_spek');
-
         if ($this->db->affected_rows() > 0) {
             return true; 
         } else {
@@ -84,28 +97,50 @@ class Kamar_model extends CI_Model {
         }
     }
 
+    public function insert_kamar($type_kamar, $price, $deskripsi, $spek, $fotos) {
+        $this->db->trans_start();
+        $kamar_data = [
+            'type_kamar' => $type_kamar,
+            'price' => $price,
+            'deskripsi' => $deskripsi
+        ];
+        $this->db->insert('kamar', $kamar_data);
+        if (!empty($spek) && is_array($spek)) {
+            foreach ($spek as $value) {
+                if (!empty(trim($value))) {
+                    $this->insert_spek($type_kamar, trim($value));
+                }
+            }
+        }
+        if (!empty($fotos['name'][0])) {
+            $this->upload_foto_kamar($type_kamar, $fotos);
+        }
+        $this->db->trans_complete(); 
+        return $this->db->trans_status();
+    }
+    
     public function insert_spek($type_kamar, $value){
         $data_spek = [
-            'type_kamar' => $type_kamar,  // Relasikan dengan ID kamar
+            'type_kamar' => $type_kamar,
             'spek' => $value
         ];
         $this->db->insert('kamar_spek', $data_spek);
     }
 
-    public function delete_kamar($type_kamar)
-    {
+    public function hapus_kamar($type_kamar){
+        $fotos = $this->db->get_where('kamar_foto', ['type_kamar' => $type_kamar])->result();
+        foreach ($fotos as $foto) {
+            $path = FCPATH . 'assets/img/upload/kamar/' . $foto->foto;
+            if (file_exists($path)) {
+                unlink($path);
+            }
+        }
         $this->db->where('type_kamar', $type_kamar);
         $this->db->delete('kamar');
-
-        if ($this->db->affected_rows() > 0) {
-            return true; 
-        } else {
-            return false;
-        }
+        redirect('admin/kamar');
     }
-
-    public function edit_kamar($type_kamar_before, $type_kamar, $price, $deskripsi)
-    {
+    
+    public function edit_kamar($type_kamar_before, $type_kamar, $price, $deskripsi){
         $data = array(
             'type_kamar' => $type_kamar,
             'price' => $price,
@@ -115,13 +150,71 @@ class Kamar_model extends CI_Model {
         return $this->db->update('kamar', $data);
     }
 
-    public function edit_foto($id_foto, $file_name)
-    {
-        $data = array(
-            'foto' => $file_name
-        );
-        $this->db->where('id_foto', $id_foto);
-        return $this->db->update('kamar_foto', $data);
+    public function update_spek($type_kamar, $spek) {
+        $this->delete_spek($type_kamar);
+        foreach ($spek as $value) {
+            $this->insert_spek($type_kamar, $value);
+        }
     }
 
+    public function edit_foto_kamar() {
+        $id_foto = $this->input->post('id_foto');
+        if (empty($_FILES['foto']['name'])) {
+            return ['status' => false, 'message' => 'Tidak ada file yang dipilih.'];
+        }
+        $foto_lama = $this->db->get_where('kamar_foto', ['id_foto' => $id_foto])->row_array();
+        if ($foto_lama && file_exists('./assets/img/upload/kamar/' . $foto_lama['foto'])) {
+            unlink('./assets/img/upload/kamar/' . $foto_lama['foto']);
+        }
+        $config = [
+            'upload_path'   => './assets/img/upload/kamar',
+            'allowed_types' => 'jpg|jpeg|png',
+            'max_size'      => 2048,
+        ];
+        $this->load->library('upload', $config);
+        $_FILES['file'] = $_FILES['foto'];
+        if (!$this->upload->do_upload('file')) {
+            return ['status' => false, 'message' => 'Upload gagal: ' . $this->upload->display_errors('', '')];
+        }
+        $file_data = $this->upload->data();
+        $updated = $this->db->update('kamar_foto', ['foto' => $file_data['file_name']], ['id_foto' => $id_foto]);
+        if ($updated) {
+            return ['status' => true, 'message' => 'Foto berhasil diedit.'];
+        } else {
+            return ['status' => false, 'message' => 'Gagal menyimpan ke database.'];
+        }
+    }
+
+    public function upload_foto_kamar($type_kamar) {
+        $config['upload_path'] = './assets/img/upload/kamar';
+        $config['allowed_types'] = 'jpg|png|jpeg';
+        $config['max_size'] = 2048;
+        $this->load->library('upload');
+        $response = ['status' => false, 'message' => ''];
+        if (!empty($_FILES['foto']['name'][0])) {
+            $filesCount = count($_FILES['foto']['name']);
+            for ($i = 0; $i < $filesCount; $i++) {
+                $_FILES['file']['name'] = $_FILES['foto']['name'][$i];
+                $_FILES['file']['type'] = $_FILES['foto']['type'][$i];
+                $_FILES['file']['tmp_name'] = $_FILES['foto']['tmp_name'][$i];
+                $_FILES['file']['error'] = $_FILES['foto']['error'][$i];
+                $_FILES['file']['size'] = $_FILES['foto']['size'][$i];
+                $this->upload->initialize($config);
+                if ($this->upload->do_upload('file')) {
+                    $fileData = $this->upload->data();
+                    $this->db->insert('kamar_foto', [
+                        'type_kamar' => $type_kamar,
+                        'foto' => $fileData['file_name']
+                    ]);
+                } else {
+                    $response['message'] = $this->upload->display_errors();
+                    return $response; 
+                }
+            }
+            $response['status'] = true;
+        } else {
+            $response['message'] = 'Tidak ada file yang diupload.';
+        }
+        return $response;
+    }
 }
